Do some major refactoring

+++ src/main/java/com/jrinehuls/foodservice/repository/NutritionFactRepository.java
@@ -3,5 +3,5 @@ package com.jrinehuls.foodservice.repository;
 import com.jrinehuls.foodservice.model.entity.NutritionFact;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NutritionFactRepository extends JpaRepository<NutritionFact, Integer> {
+public interface NutritionFactRepository extends JpaRepository<NutritionFact, Long> {
 }
